consolidate all api requests into one place
