--- src/sendInit.ts
@@ -1,21 +1,0 @@
-import { config } from './config';
-import { sendToAPI } from './api';
-
-export const sendInitToAPI = async () => {
-  if (config.generateOnly) {
-    return;
-  }
-
-  const [repoOwner, repoName] = config.repository.split('/');
-
-  return sendToAPI({
-    action: 'init',
-    payload: {
-      projectId: config.lostPixelProjectId,
-      branchName: config.commitRefName,
-      repoOwner,
-      repoName,
-      commit: config.commitHash,
-    },
-  });
-};
